seems like it works

+++ lib/puppet/provider/windows_env/windows_env.rb
@@ -111,46 +111,7 @@ Puppet::Type.type(:windows_env).provide(:windows_env) do
   end
 
   def exists?
-    # For testing registry open result
-    _ERROR_FILE_NOT_FOUND = 2
-
-    if @resource[:user]
-      @reg_hive = Win32::Registry::HKEY_USERS
-      @user_sid = name_to_sid(@resource[:user])
-      @user_sid or self.fail "Username '#{@resource[:user]}' could not be converted to a valid SID"
-      @reg_path = "#{@user_sid}\\Environment"
-
-      begin
-        @reg_hive.open(@reg_path) {}
-      rescue Win32::Registry::Error => error
-        if error.code == _ERROR_FILE_NOT_FOUND
-          load_user_hive
-        else
-          reg_fail("Can't access Environment for user '#{@resource[:user]}'. Opening", error)
-        end
-      end
-    else
-      @reg_hive = Win32::Registry::HKEY_LOCAL_MACHINE
-      @reg_path = 'System\CurrentControlSet\Control\Session Manager\Environment'
-    end
-
-    @sep = @resource[:separator]
-
-    @reg_types = { :REG_SZ => Win32::Registry::REG_SZ, :REG_EXPAND_SZ => Win32::Registry::REG_EXPAND_SZ }
-    @reg_type = @reg_types[@resource[:type]]
-
-    begin
-      # key.read returns '[type, data]' and must be used instead of [] because [] expands %variables%. 
-      @reg_hive.open(@reg_path) { |key| @value = key.read(@resource[:variable])[1] } 
-    rescue Win32::Registry::Error => error
-      if error.code == _ERROR_FILE_NOT_FOUND
-        debug "Environment variable #{@resource[:variable]} not found"
-        return false
-      end
-      reg_fail('reading', error)
-    end
-
-    @value = @value.split(@sep)
+    @value = get_current_reg_val()
 
     # Assume that if the user says 'ensure => absent', they want the value to
     # be removed regardless of its position, i.e. use the 'insert' behavior
@@ -161,25 +122,23 @@ Puppet::Type.type(:windows_env).provide(:windows_env) do
       @resource[:mergemode] = :insert
     end
 
-    case @resource[:mergemode]
-    when :clobber
-      # When 'ensure == absent' in clobber mode, we delete the variable itself, regardless of its content, so
-      # don't bother checking the content in this case. 
-      @resource[:ensure] == :present ? @value == @resource[:value] : true
-    when :insert
-      # FIXME: this is a weird way to do this
-      # verify all elements are present and they appear in the correct order
-      indexes = @resource[:value].map { |x| @value.find_index { |y| x.casecmp(y) == 0 } }
-      if indexes.count == 1
-        indexes == [nil] ? false : true
-      else
-        indexes.each_cons(2).all? { |a, b| a && b && a < b }
+    # We might be running in a stale environment and want to update ruby's
+    # environment to match what's in the registry. While we could return a
+    # value from #exists? in this case that forces all the other code to run,
+    # this would probably not be what the user wants (since it'll trigger
+    # subscribed resources and such even when the registry was already
+    # correct). So instead, just update the environment quietly.
+    if @resource[:update_ruby_env]
+      exists_current_env = exists_helper(ENV[@resource[:variable]])
+      if (@resource[:ensure] == :present && ! exists_current_env) ||
+         (@resource[:ensure] == :absent && exists_current_env)
+        # TODO: utf-16le stuff eith ENV ???
+        debug "updating ruby envrionment to match puppet defined value"
+        ENV[@resource[:variable]] = get_desired_value(ENV[@resource[:variable]])
       end
-    when :append
-      @value.map { |x| x.downcase }[(-1 * @resource[:value].count)..-1] == @resource[:value].map { |x| x.downcase }
-    when :prepend
-      @value.map { |x| x.downcase }[0..(@resource[:value].count - 1)] == @resource[:value].map { |x| x.downcase }
     end
+
+    exists_helper(@value)
   end
 
   def create
@@ -192,51 +151,38 @@ Puppet::Type.type(:windows_env).provide(:windows_env) do
       @resource[:mergemode] = :clobber
     end
 
-    case @resource[:mergemode]
-    when :clobber
+    desired_value = get_desired_value(@value)
+
+    # Creating a new key or updating an old?
+    if @value.nil?
       @reg_type = Win32::Registry::REG_SZ unless @reg_type
       begin
         @reg_hive.create(@reg_path, Win32::Registry::KEY_ALL_ACCESS | Win32::Registry::KEY_WOW64_64KEY) do |key| 
-          val = @resource[:value].join(@sep)
-          key[@resource[:variable], @reg_type] = val
-          if @resource[:update_ruby_env] == :true
-            ENV[@resource[:variable]] = val
-          end
+          key[@resource[:variable], @reg_type] = desired_value
         end
       rescue Win32::Registry::Error => error
         reg_fail('creating', error)
       end
-    # the position at which the new value will be inserted when using insert is
-    # arbitrary, so may as well group it with append.
-    when :insert, :append
-      # delete if already in the string and move to end.
-      remove_value
-      @value = @value.concat(@resource[:value])
-      key_write
-    when :prepend
-      # delete if already in the string and move to front
-      remove_value
-      @value = @resource[:value].concat(@value)
+    else
+      @value = desired_value
       key_write
     end
+
     broadcast_changes
   end
 
   def destroy
     debug "Removing value from environment variable '#{@resource[:variable]}', or removing variable itself"
-    case @resource[:mergemode]
-    when :clobber
+
+    @value = get_desired_value(@value)
+    if @value.nil?
       key_write do |key|
         self.class::WinAPI.delete_value(key, @resource[:variable])
-        # XXX: you cannot pass UTF-16LE strings to ENV.delete or ENV[]
-        if @resource[:update_ruby_env] == :true
-          ENV[@resource[:variable].encode("UTF-8")] = nil
-        end
       end
-    when :insert, :append, :prepend
-      remove_value
+    else
       key_write
     end
+
     broadcast_changes
   end
 
@@ -248,7 +194,7 @@ Puppet::Type.type(:windows_env).provide(:windows_env) do
 
   def type=(newtype)
     newtype = @reg_types[newtype]
-    key_write { |key| key[@resource[:variable], newtype] = @value.join(@sep) }
+    key_write { |key| key[@resource[:variable], newtype] = @value }
     broadcast_changes
   end
 
@@ -278,8 +224,119 @@ Puppet::Type.type(:windows_env).provide(:windows_env) do
     self.fail "#{action} '#{@reg_hive.name}:\\#{@reg_path}\\#{@resource[:variable]}' returned error #{error.code}: #{error.message}"
   end
 
-  def remove_value
-    @value = @value.delete_if { |x| @resource[:value].find { |y| y.casecmp(x) == 0 } }
+  def remove_value(current)
+    current.delete_if { |x| @resource[:value].find { |y| y.casecmp(x) == 0 } }
+    current
+  end
+
+  # Determine if #exists? should return true. This is abstracted out because we need to
+  # check two kinds of existence: existence in the registry and existence in the current
+  # ruby process' environment.
+  # current_value may be nil if the variable doesn't exist.
+  def exists_helper(current_value)
+    current_value = current_value.nil? ? nil : current_value.split(@resource[:separator])
+
+    case @resource[:mergemode]
+    when :clobber
+      # When 'ensure == absent' in clobber mode, we delete the variable itself, regardless of its content, so
+      # don't bother checking the content in this case, just whether the variable exists or not.
+      if @resource[:ensure] == :absent
+        current_value.nil?
+      else
+        # TODO: remove
+        debug "current_value = '#{current_value}' @resource[:value] == '#{@resource[:value]}'"
+        current_value == @resource[:value]
+      end
+    when :insert
+      # FIXME: this is a weird way to do this
+      # verify all elements are present and they appear in the correct order
+      indexes = @resource[:value].map { |x| current_value.find_index { |y| x.casecmp(y) == 0 } }
+      if indexes.count == 1
+        indexes == [nil] ? false : true
+      else
+        indexes.each_cons(2).all? { |a, b| a && b && a < b }
+      end
+    when :append
+      current_value.map { |x| x.downcase }[(-1 * @resource[:value].count)..-1] == @resource[:value].map { |x| x.downcase }
+    when :prepend
+      current_value.map { |x| x.downcase }[0..(@resource[:value].count - 1)] == @resource[:value].map { |x| x.downcase }
+    end
+  end
+
+  # Get the current value of the environment variable in the registry. Return
+  # nil if the variable is not present.
+  def get_current_reg_val
+    # For testing registry open result
+    _ERROR_FILE_NOT_FOUND = 2
+
+    if @resource[:user]
+      @reg_hive = Win32::Registry::HKEY_USERS
+      user_sid = name_to_sid(@resource[:user])
+      user_sid or self.fail "Username '#{@resource[:user]}' could not be converted to a valid SID"
+      @reg_path = "#{user_sid}\\Environment"
+
+      begin
+        @reg_hive.open(@reg_path) {}
+      rescue Win32::Registry::Error => error
+        if error.code == _ERROR_FILE_NOT_FOUND
+          load_user_hive(user_sid)
+        else
+          reg_fail("Can't access Environment for user '#{@resource[:user]}'. Opening", error)
+        end
+      end
+    else
+      @reg_hive = Win32::Registry::HKEY_LOCAL_MACHINE
+      @reg_path = 'System\CurrentControlSet\Control\Session Manager\Environment'
+    end
+
+    @reg_types = { :REG_SZ => Win32::Registry::REG_SZ, :REG_EXPAND_SZ => Win32::Registry::REG_EXPAND_SZ }
+    @reg_type = @reg_types[@resource[:type]]
+
+    current_value = nil
+    begin
+      # key.read returns '[type, data]' and must be used instead of [] because [] expands %variables%. 
+      @reg_hive.open(@reg_path) { |key| current_value = key.read(@resource[:variable])[1] } 
+    rescue Win32::Registry::Error => error
+      if error.code == _ERROR_FILE_NOT_FOUND
+        debug "Environment variable #{@resource[:variable]} not found"
+      else
+        reg_fail('reading', error)
+      end
+    end
+
+    current_value
+  end
+
+  # Get the desired value of the environment variable. If
+  # the variable is to be deleted, return nil.
+  def get_desired_value(current_value)
+    current_value = current_value.nil? ? nil : current_value.split(@resource[:separator])
+
+    if @resource[:ensure] == :present
+      case @resource[:mergemode]
+      when :clobber
+        val = [@resource[:value]]
+      # the position at which the new value will be inserted when using insert is
+      # arbitrary, so may as well group it with append.
+      when :insert, :append
+        # delete if already in the string and move to end.
+        val = remove_value(current_value)
+        val = val.concat(@resource[:value])
+      when :prepend
+        # delete if already in the string and move to front
+        val = remove_value(current_value)
+        @resource[:value].concat(val)
+      end
+    else
+      case @resource[:mergemode]
+      when :clobber
+        val = nil
+      when :insert, :append, :prepend
+        val = remove_value(current_value)
+      end
+    end
+
+    val.nil? ? nil : val.join(@resource[:separator])
   end
 
   def key_write(&block)
@@ -292,13 +349,7 @@ Puppet::Type.type(:windows_env).provide(:windows_env) do
       else
         newtype = @reg_types[self.type]
       end
-        block = proc do |key|
-          val = @value.join(@sep)
-          key[@resource[:variable], newtype] = val
-          if @resource[:update_ruby_env] == :true
-            ENV[@resource[:variable]] = val
-          end
-        end
+      block = proc { |key| key[@resource[:variable], newtype] = @value }
     end
     @reg_hive.open(@reg_path, Win32::Registry::KEY_WRITE | Win32::Registry::KEY_WOW64_64KEY, &block) 
   rescue Win32::Registry::Error => error
@@ -321,12 +372,12 @@ Puppet::Type.type(:windows_env).provide(:windows_env) do
 
   # This is the best solution I found to (at least mostly) reliably locate a user's 
   # ntuser.dat: http://stackoverflow.com/questions/1059460/shgetfolderpath-for-a-specific-user
-  def load_user_hive
+  def load_user_hive(user_sid)
     debug "Loading NTUSER.DAT for '#{@resource[:user]}'"
 
     home_path = nil
     begin
-      Win32::Registry::HKEY_LOCAL_MACHINE.open("SOFTWARE\\Microsoft\\Windows NT\\CurrentVersion\\ProfileList\\#{@user_sid}") do |key|
+      Win32::Registry::HKEY_LOCAL_MACHINE.open("SOFTWARE\\Microsoft\\Windows NT\\CurrentVersion\\ProfileList\\#{user_sid}") do |key|
         home_path = key['ProfileImagePath']
       end
     rescue Win32::Registry::Error => error
@@ -336,13 +387,13 @@ Puppet::Type.type(:windows_env).provide(:windows_env) do
     ntuser_path = File.join(home_path, 'NTUSER.DAT')
 
     Puppet::Util::Windows::Security.with_privilege(Puppet::Util::Windows::Security::SE_RESTORE_NAME) do
-      result = self.class::WinAPI.RegLoadKey(Win32::Registry::HKEY_USERS.hkey, @user_sid, ntuser_path)
+      result = self.class::WinAPI.RegLoadKey(Win32::Registry::HKEY_USERS.hkey, user_sid, ntuser_path)
       unless result == 0
         raise Puppet::Util::Windows::Error.new("Could not load registry hive for user '#{@resource[:user]}'", result)
       end
     end
 
-    self.class.loaded_hives << { :user_sid => @user_sid, :username => @resource[:user] }
+    self.class.loaded_hives << { :user_sid => user_sid, :username => @resource[:user] }
   end
 end
